izumi

+++ project/src/commands/user/RemoveAccountCommand.java
@@ -1,5 +1,8 @@
 package commands.user;
 
+import java.util.HashMap;
+import java.util.Map;
+
 import bean.User;
 import commands.AbstractCommand;
 import context.RequestContext;
@@ -19,10 +22,21 @@ public class RemoveAccountCommand extends AbstractCommand {
 
 		AbstractDaoFactory factory=AbstractDaoFactory.getFactory();
 		UserDao dao=factory.getUserDao();
-		dao.removeUser(user.getUserIdentifiedName(),user.getUserMail(),user.getUserPassword());
-		System.out.println("ユーザーの削除(unused)");
 
-		SessionManager.invalidate();
+		String inputPassword = reqc.getParameter("userPassword")[0];
+
+		if(inputPassword == user.getUserPassword()) {
+			dao.removeUser(user.getUserIdentifiedName(),user.getUserMail(),user.getUserPassword());
+			System.out.println("ユーザーの削除(unused)");
+			SessionManager.invalidate();
+		}else {
+			Map result = new HashMap();
+			result.put("ngpass", "true");
+			result.put("duplicateCheck", "false");
+			resc.setResult(result);
+			resc.setTarget("users/settings");
+			return resc;
+		}
 
 		resc.setTarget("index");
